MCP/requirements.txt
streamlit>=1.29.0
langchain>=0.1.14
langchain-core>=0.1.37
langchain-ollama>=0.1.3
PyPDF2>=3.0.1
python-dotenv>=1.0.1
fastmcp
streamlit>=1.32.0
aiohttp>=3.9.3
